fix : 코디 추가 변경 완료

See merge request s06-webmobile2-sub2/S06P12B108!43

+++ backend/src/main/java/com/web/gdup/domain/cody/dto/ClothingInCody.java
@@ -10,7 +10,7 @@ import lombok.*;
 public class ClothingInCody {
 
     private int clothingId;
-    private int m;
+    private double m;
     private int x;
     private int y;
     private int z;
