+++ backend/src/main/java/com/web/gdup/domain/cody/entity/CodyClothingInfo.java
@@ -29,6 +29,6 @@ public class CodyClothingInfo {
     private int x;
     private int y;
     private int z;
-    private int m;
+    private double m;
 
 }
